fix(contract): Gemini レビュー指摘の対応 (Stage 10-A)
- Issue 1 (HIGH): SUPPORTED_VERSIONS 配列を導入、verifyEnvelope は受信
  signature を全 supported version で検証試行。非アトミック OTA migration
  に耐える構造に。verifyEnvelope の引数から version を撤去 (内部試行)。
  EnvelopeVerifyResult.ok=true 時に合致した version を返す。
- Issue 2 (LOW): crypto.subtle.verify 採用、constantTimeEquals + 文字列
  比較を撤去。length early return のタイミングリーク懸念解消。不正 hex
  は signature_mismatch として扱う。
- Issue 4 (防御): validateLogRow に Array.isArray guard 追加。エラー型
  に shape:is_array / shape:not_an_object を追加。
- Issue 3 (準備): emit-c-header.ts の helper 引数を unsigned long long
  に変更 + 2038 言及コメント。yaml 側の cast 揃えは Stage 10-B。

contract.generated.h と docs/CONTRACT.md は npm run codegen で再生成。
007_contract_functions.sql は payload schema 不変ゆえ実質変更なし。

+++ contract/codegen/emit-c-header.ts
@@ -38,7 +38,12 @@ constexpr int YAKIIMO_CHANNELS_COUNT = ${CHANNELS.length};
 
 // 署名対象メッセージを構築する。
 // 形式: \`\${version}\\n\${timestamp_unix_sec}\\n\${body}\`
-inline std::string yakiimo_build_signed_message(unsigned long timestamp_unix_sec, const std::string& body) {
+//
+// 注: 引数は unsigned long long (64-bit)。ESP32 firmware (Arduino-ESP32) の
+// time_t は signed 32-bit ゆえ 2038-01-19 にラップする (詳細は ADR-0002)。
+// ここでは内部一貫性 (header timestamp と HMAC message timestamp の cast 幅
+// 統一) のみ確保し、根治は framework 64-bit time_t 採用待ち。
+inline std::string yakiimo_build_signed_message(unsigned long long timestamp_unix_sec, const std::string& body) {
   return std::to_string(YAKIIMO_CONTRACT_VERSION) + "\\n" + std::to_string(timestamp_unix_sec) + "\\n" + body;
 }
 `;
